# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GitHub Stats API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ContributionDay {
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Number of contributions on this day */
  count: number;
  /** Contribution intensity level (0-4) */
  level: number;
}

export interface ContributionWeek {
  days: ContributionDay[];
}

export interface ContributionCalendar {
  /** Total contributions in the requested period */
  totalContributions: number;
  weeks: ContributionWeek[];
  /** Year of this contribution data */
  year: number;
}

export interface GithubStats {
  username: string;
  /** Total contributions across all time (all years from account creation to now) */
  totalContributions: number;
  /** Current consecutive days with contributions (up to and including today) */
  currentStreak: number;
  /** Longest consecutive days streak ever */
  longestStreak: number;
  /**
   * Start date of current streak (YYYY-MM-DD)
   * @nullable
   */
  currentStreakStart: string | null;
  /**
   * End date of current streak (YYYY-MM-DD)
   * @nullable
   */
  currentStreakEnd: string | null;
  /**
   * Start date of longest streak (YYYY-MM-DD)
   * @nullable
   */
  longestStreakStart: string | null;
  /**
   * End date of longest streak (YYYY-MM-DD)
   * @nullable
   */
  longestStreakEnd: string | null;
  /** Total contributions in the current year */
  contributionsThisYear: number;
  /** Date the GitHub account was created (YYYY-MM-DD) */
  accountCreatedAt: string;
  /**
   * Date of first ever contribution (YYYY-MM-DD)
   * @nullable
   */
  firstContributionDate: string | null;
}

export interface GithubProfile {
  username: string;
  /** @nullable */
  name: string | null;
  /** @nullable */
  bio: string | null;
  avatarUrl: string;
  followers: number;
  following: number;
  publicRepos: number;
  publicGists: number;
  createdAt: string;
  /** @nullable */
  location: string | null;
  /** @nullable */
  blog: string | null;
  /** @nullable */
  company: string | null;
  /** @nullable */
  twitterUsername: string | null;
}

export type GetGithubStatsParams = {
  /**
   * Optional GitHub personal access token for higher rate limits
   */
  token?: string;
};

export type GetGithubContributionsParams = {
  /**
   * Optional GitHub personal access token
   */
  token?: string;
  /**
   * Year to fetch contributions for (defaults to current year)
   */
  year?: number;
};

export type GetGithubProfileParams = {
  /**
   * Optional GitHub personal access token
   */
  token?: string;
};