# lib/api-zod/src/generated/types/contributionCalendar.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GitHub Stats API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContributionWeek } from "./contributionWeek";

export interface ContributionCalendar {
  /** Total contributions in the requested period */
  totalContributions: number;
  weeks: ContributionWeek[];
  /** Year of this contribution data */
  year: number;
}