# lib/api-zod/src/generated/types/contributionDay.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GitHub Stats API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContributionDay {
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Number of contributions on this day */
  count: number;
  /** Contribution intensity level (0-4) */
  level: number;
}