# lib/api-zod/src/generated/types/contributionWeek.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GitHub Stats API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContributionDay } from "./contributionDay";

export interface ContributionWeek {
  days: ContributionDay[];
}