# lib/api-zod/src/generated/types/githubProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GitHub Stats API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubProfile {
  username: string;
  /** @nullable */
  name: string | null;
  /** @nullable */
  bio: string | null;
  avatarUrl: string;
  followers: number;
  following: number;
  publicRepos: number;
  publicGists: number;
  createdAt: string;
  /** @nullable */
  location: string | null;
  /** @nullable */
  blog: string | null;
  /** @nullable */
  company: string | null;
  /** @nullable */
  twitterUsername: string | null;
}