# lib/api-zod/src/generated/types/githubStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GitHub Stats API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubStats {
  username: string;
  /** Total contributions across all time (all years from account creation to now) */
  totalContributions: number;
  /** Current consecutive days with contributions (up to and including today) */
  currentStreak: number;
  /** Longest consecutive days streak ever */
  longestStreak: number;
  /**
   * Start date of current streak (YYYY-MM-DD)
   * @nullable
   */
  currentStreakStart: string | null;
  /**
   * End date of current streak (YYYY-MM-DD)
   * @nullable
   */
  currentStreakEnd: string | null;
  /**
   * Start date of longest streak (YYYY-MM-DD)
   * @nullable
   */
  longestStreakStart: string | null;
  /**
   * End date of longest streak (YYYY-MM-DD)
   * @nullable
   */
  longestStreakEnd: string | null;
  /** Total contributions in the current year */
  contributionsThisYear: number;
  /** Date the GitHub account was created (YYYY-MM-DD) */
  accountCreatedAt: string;
  /**
   * Date of first ever contribution (YYYY-MM-DD)
   * @nullable
   */
  firstContributionDate: string | null;
}